crop/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.26.3
pytz==2023.3.post1
scikit-learn==1.2.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.4
